fix: lifetime errors
