+++ llama-rs/src/common/model.rs
@@ -3,26 +3,11 @@ use super::inference::{
     InferenceSnapshot, SnapshotError,
 };
 use super::load::{LoadError, LoadProgress};
-use super::token::{TokenBias, TokenId};
+use super::token::TokenId;
 use super::vocabulary::Vocabulary;
-use crate::ggml;
-use crate::mulf;
 use partial_sort::PartialSort;
 use rand::{distributions::WeightedIndex, prelude::Distribution};
-
 use std::path::Path;
-
-#[derive(Debug, Default, PartialEq, Eq, PartialOrd, Ord)]
-pub struct Hyperparameters {
-    pub n_vocab: i32,
-    pub n_ctx: i32,
-    pub n_embd: i32,
-    pub n_mult: i32,
-    pub n_head: i32,
-    pub n_layer: i32,
-    pub n_rot: Option<i32>,
-    pub f16_: i32,
-}
 
 /// Used in a call to `evaluate` to request information from the transformer.
 #[derive(Default)]
@@ -36,62 +21,21 @@ pub struct EvaluateOutputRequest {
 }
 
 pub trait Model {
-    type OutputType;
+    type Weights;
+    type HP;
 
-    fn load(
+
+    fn load<'a>(
         &self,
         path: impl AsRef<Path>,
         n_ctx: i32,
-        load_progress_callback: impl Fn(LoadProgress<T>),
-    ) -> Result<(Self::OutputType, Vocabulary), LoadError>;
+        load_progress_callback: impl Fn(LoadProgress<'a, Self::HP>),
+    ) -> Result<(Self::Weights, Vocabulary), LoadError>
+    where
+        Self::HP: 'a;
 
     /// Starts a new `InferenceSession` for this model.
-    fn start_session(&self, params: InferenceSessionParameters) -> InferenceSession {
-        let Hyperparameters {
-            n_ctx,
-            n_embd,
-            n_layer,
-            n_vocab,
-            ..
-        } = self.hparams;
-
-        let ctx_size = {
-            let mut ctx_size = 0;
-            ctx_size += mulf!(
-                n_ctx,
-                n_layer,
-                n_embd,
-                ggml::type_sizef(params.memory_k_type.into())
-            ); // memory_k
-            ctx_size += mulf!(
-                n_ctx,
-                n_layer,
-                n_embd,
-                ggml::type_sizef(params.memory_v_type.into())
-            ); // memory_v
-            ctx_size += (5 + 10 * n_layer as u64) * 256; // object overhead
-            ctx_size
-        };
-
-        let session_ctx = ggml::Context::init(ctx_size as usize);
-
-        // Initialize key + value memory tensors
-        let n_mem = n_layer * n_ctx;
-        let n_elements = n_embd * n_mem;
-        let memory_k = session_ctx.new_tensor_1d(params.memory_k_type.into(), n_elements);
-        let memory_v = session_ctx.new_tensor_1d(params.memory_v_type.into(), n_elements);
-
-        InferenceSession {
-            _session_ctx: session_ctx,
-            params,
-            memory_k,
-            memory_v,
-            n_past: 0,
-            mem_per_token: 0,
-            tokens: vec![],
-            last_logits: vec![0.0; n_vocab as usize],
-        }
-    }
+    fn start_session(&self, params: InferenceSessionParameters) -> InferenceSession;
 
     fn sample_top_p_top_k(
         &self,
